Repo Reorg: consolidate validation modules
