+++ src/nisarqa/validate/xml_spec_checker/h5_parser.py
