+++ src/nisarqa/validate/xml_spec_checker/xml_parser.py
